linux libcurl issue fixed

+++ example/create_invoice.php
@@ -6,7 +6,7 @@ $data = array(
     'name' => 'TIP',
     'desc' => @$_GET['val'] . ' DERO',
 	'ttl' => 1200,
-	'price' => (int) @$_GET['val'] * 1000000000000,
+	'price' => (float) @$_GET['val'] * 1000000000000,
 	'webhook_data' => [
 		'customer_nick' => @$_GET['nick'],
 	]
